Import from Aws(annotation only)
